Address potential webview startup race condition
This might be paranoia from me but it is theoretically possible.

Your code here assumes that when environment_completed.handler is called, the message loop has finished initializing the thread is waiting for messages.

But you don't have any guarantee that this is true. It all depends on how the microsoft library invokes the environment_completed.handler and later on, the controller_completed.handler. It is theoretically possible for line 238 to be called AFTER the WV library has invoked both the environment_completed.handler and the controller_completed.handler callback functions. If this happens, then all PostThreadMessage messages in those callback functions will get dropped and this would be a bug.

This may work 99% of the time and then fail randomly due to how timing dependent it is. It might also never show up on our machines but could show up on some random user's machine.

The safest thing to do is to make your callbacks wait until the message loop has begun spinning. My proposed change does this the easiest way I can think of.

Thoughts?